# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gratification API",
        "description": "Documentação da API do Gratification",
        "contact": {
            "email": "suporte@gratification.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/profile/update": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Atualizar as informações do perfil do usuário",
                "operationId": "6ed305f4c4bd48cdaa2e5066f5b7f156",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "joao.silva@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Perfil atualizado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profile/destroy": {
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Excluir a conta do usuário",
                "operationId": "f0b6b3203be521955bd3605d26943b51",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "senhaatual"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Conta excluída com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Conta excluída com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}